리눅스 서버 업로드 준비

+++ src/main/java/com/upbit/market/MarketDto.java
@@ -169,7 +169,7 @@ public class MarketDto {
     /**
      * count
      */
-    int buyCount = 0;
+    Double buyCount = 0.0;
 //
 //    /**
 //     * 매도호가
